Version 1.0.0
Model and functions to retrieve data is built in the Stock_predictor_Source.py file well all functions for training and saving the model are in the Stock_Predictor_main.py file.

Functions for displaying training and testing plots was also implemented.

Model is over-fitting the data fine-tuning will be the next-step.